Restrict visibility of exported symbols
By default, GCC makes every symbol in the .so visible. However, to avoid clashes
and maintain a clear API it is better to export only a whitelist.

This patch does two things and a hack :

  - set the default visibility to "hidden".
  - annotate some exported functions to "default", ie. externally visible.

  - (the hack) adjust src/glyrc/autohelp.c so that it links directly against
    stringlib objects (instead of linking through libglyr)

The last point is certainly questionable. The problem is that a few functions
from stringlib are referenced in autohelp.c. An alternative solution could be to
rename get_next_word into glyr_get_next_word and make it part of the public API,
but that's not ideal either.