Use BUILD_DATE / BUILD_TIME instead of cpp macros
`glyrc --version` returns the date and time the library was built. To do that,
it uses the standard macros `__DATE__` and `__TIME__`.

The problem with this is that it makes the build not reproducible, since a
binary produced a few seconds later will obviously be different.

This adds a mechanism to bypass this: if the environment variables `BUILD_DATE`
and `BUILD_TIME` are exported at compile-time, they will be used instead.

This is part of an effort to make Debian packages reproducible, but should
benefit other distributions. More information can be found at [1], and [2] for
this particular issue.

https://wiki.debian.org/ReproducibleBuilds/About
https://wiki.debian.org/ReproducibleBuilds/TimestampsFromCPPMacros